Fix "cell is slipping away under finger"-bug by rounding `distance` to the an integer.
Setting a non-integral value to `self.collectionView.contentOffset` will get rounded automatically anyway but `self.currentViewCenter` and `self.currentView.center` won't and that's where the values slowly diverge if the speed is not dividable without rest by 60 (the default fps). The default speed is 300 but bumping it up to e.g. 500 for faster scrolling caused the bug to happen.